refactor: déplacer isOfferActive dans ui/utils et tester le statut en attente
