+++ ui/app/(candidat)/emploi/[type]/[id]/[intitule-offre]/CandidaterButton.tsx
@@ -5,11 +5,11 @@ import Button from "@codegouvfr/react-dsfr/Button"
 import { Box } from "@mui/material"
 import { useState } from "react"
 import type { ILbaItemLbaCompanyJson, ILbaItemLbaJobJson, ILbaItemPartnerJobJson } from "shared"
-import { isOfferActive } from "@/app/(candidat)/(recherche)/recherche/_components/is-offer-active"
 import { useDisclosure } from "@/app/hooks/use-disclosure"
 import { useSubmitCandidature } from "@/components/ItemDetail/CandidatureLba/services/submit-candidature"
 import ItemDetailApplicationsStatus from "@/components/ItemDetail/ItemDetailServices/ItemDetailApplicationStatus"
 import { notifyJobPostulerV3 } from "@/utils/api"
+import { isOfferActive } from "@/utils/is-offer-active"
 import { getMatomoJobOfferType, MATOMO_EVENTS, pushMatomoEvent } from "@/utils/matomo-utils"
 import { SendPlausibleEvent } from "@/utils/plausible"
