+++ ui/components/ItemDetail/ItemDetailServices/EmployeurPresentationBlock.tsx
@@ -3,9 +3,9 @@ import { fr } from "@codegouvfr/react-dsfr"
 import { Box, Stack, Typography } from "@mui/material"
 import type { ILbaItemLbaCompanyJson, ILbaItemNaf, ILbaItemPartnerJobJson } from "shared"
 import { LBA_ITEM_TYPE } from "shared/constants/lbaitem"
-import { isOfferActive } from "@/app/(candidat)/(recherche)/recherche/_components/is-offer-active"
 import { DsfrLink } from "@/components/dsfr/DsfrLink"
 import { LbaJobEngagement } from "@/components/ItemDetail/LbaJobComponents/LbaJobEngagement"
+import { isOfferActive } from "@/utils/is-offer-active"
 import { getCompanySize } from "./get-company-size"
 import HiringCountBox from "./HiringCountBox"
 import ItemDistanceToCenter from "./ItemDistanceToCenter"
