+++ ui/components/ItemDetail/LbaJobComponents/LbaJobCfaDetail.tsx
@@ -4,7 +4,6 @@ import { Box, Link, Stack, Typography } from "@mui/material"
 import Image from "next/image"
 import { useEffect } from "react"
 import type { ILbaItemPartnerJobJson } from "shared"
-import { isOfferActive } from "@/app/(candidat)/(recherche)/recherche/_components/is-offer-active"
 import { DsfrLink } from "@/components/dsfr/DsfrLink"
 import { ContratBlock } from "@/components/ItemDetail/ItemDetailServices/ContratBlock"
 import { EmployeurPresentationBlock } from "@/components/ItemDetail/ItemDetailServices/EmployeurPresentationBlock"
@@ -13,6 +12,7 @@ import ItemLocalisation from "@/components/ItemDetail/ItemDetailServices/ItemLoc
 import { BAD_DESCRIPTION_LENGTH, JobDescription } from "@/components/ItemDetail/ItemDetailServices/JobDescription"
 import { JobPostingSchema } from "@/components/ItemDetail/JobPostingSchema"
 import { notifyJobDetailViewV3 } from "@/utils/api"
+import { isOfferActive } from "@/utils/is-offer-active"
 import { SendPlausibleEvent } from "@/utils/plausible"
 import LbaJobAcces from "./LbaJobAcces"
 import LbaJobCompetences from "./LbaJobCompetences"
