+++ ui/utils/is-offer-active.ts
